pajlada said to revert my changes
Given his feedback it's wrong to keep !debug playsound a thing if the playsound module is disabled
