+++ pajbot/modules/playsound.py
@@ -450,6 +450,29 @@ class PlaysoundModule(BaseModule):
             log_msg = f"The {playsound_name} playsound has been removed"
             AdminLogManager.add_entry("Playsound removed", source, log_msg)
 
+    @staticmethod
+    def debug_playsound_command(bot, source, message, **rest):
+        """Method for debugging (printing info about) playsounds.
+        Usage: !debug playsound PLAYSOUNDNAME
+        """
+        playsound_name = PlaysoundModule.massage_name(message.split(" ")[0])
+        # check for empty string
+        if not playsound_name:
+            bot.whisper(source, "Invalid usage. Correct syntax: !debug playsound <name>")
+            return
+
+        with DBManager.create_session_scope() as session:
+            playsound = session.query(Playsound).filter(Playsound.name == playsound_name).one_or_none()
+
+            if playsound is None:
+                bot.whisper(source, "No playsound with that name exists.")
+                return
+
+            bot.whisper(
+                source,
+                f"name={playsound.name}, link={playsound.link}, volume={playsound.volume}, cooldown={playsound.cooldown}, enabled={playsound.enabled}",
+            )
+
     def load_commands(self, **options):
         from pajbot.models.command import Command
         from pajbot.models.command import CommandExample
@@ -595,6 +618,32 @@ class PlaysoundModule(BaseModule):
                             "Remove an existing playsound",
                             chat="user:!remove playsound doot\n" "bot>user:Successfully removed your playsound",
                             description='Removes the "doot" playsound.',
+                        ).parse()
+                    ],
+                )
+            },
+        )
+
+        self.commands["debug"] = Command.multiaction_command(
+            level=100,
+            delay_all=0,
+            delay_user=0,
+            default=None,
+            command="debug",
+            commands={
+                "playsound": Command.raw_command(
+                    self.debug_playsound_command,
+                    level=250,
+                    delay_all=5,
+                    delay_user=15,
+                    description="Prints data about a playsound",
+                    examples=[
+                        CommandExample(
+                            None,
+                            'Get information about the "doot" playsound',
+                            chat="user:!debug playsound doot\n"
+                            "bot>user: name=doot, link=https://playsound.pajbot.com/common/doot.ogg, volume=100, "
+                            "cooldown=None, enabled=True",
                         ).parse()
                     ],
                 )
